requirements.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
schedule>=1.2.1
